third_party/blink/web_tests/http/tests/misc/window-open-with-additional-bool-parameter.html

<html>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.setPopupBlockingEnabled(false);
    testRunner.dumpAsText();
}

try {
    window.open('about:blank', "_blank", "", true);
    document.write("success");
} catch (ex) {
    document.write("FAIL - caught exception: " + ex);
}

if (window.testRunner) {
    testRunner.notifyDone();
}

</script>
<div>This tests that calling window.open() with a boolean as a fourth parameter does no throw an exception.</div>
</html>
